Add parsing of json response
This commit pulls a new library from the FixedIT library zoo that
is then used to parse the json response from the server.

The json-glib library is prebuilt by FixedIT Consulting for the
different Axis cameras and several different versions of the AXIS
ACAP SDKs. This makes it as easy as a few lines of code to use the
library in the application and maintenance is minimal.